refactor(rust-api): extract recipe-preset CRUD/catalog/validation into recipe_presets.rs (sc-1639 Phase 1, slice 4 / final)
Final behavior-preserving slice of the sc-1639 lib.rs module split. Moves the
recipe-preset-exclusive code out of lib.rs into a new recipe_presets.rs (~960
lines, 50 items):
- handlers: list/get/create/update/delete/duplicate_recipe_preset
- catalog: recipe_preset_catalog, recipe_preset_scope_order
- normalize/finalize + workflow defaults + scope/context/write-location helpers
- validators: validate_recipe_preset_{id,workflow,order,defaults,model_workflow,
  lora_compatibility,prompt} + validate_required_string_field + normalize_*_for_write
- preset lora serialization: recipe_preset_loras, preset_prompt, preset_lora_id,
  preset_lora_weight, serialize_preset_lora, normalize_recipe_preset_loras
- id/name dedup: slugify_preset_id, next_duplicate_preset_{id,name}, *_exists

Shared helpers stay at root: the job-lora serialization used by loras.rs
(job_lora_id, serialize_job_lora, normalize_inline_job_lora, preferred_lora_*),
catalog_delete_warnings/job_references_catalog_item (model+lora delete), and
project_path_for_id + the family validators/util layer. recipe_presets.rs
reaches them via `use super::*`; same pub(crate) + `use recipe_presets::*`
re-export pattern keeps generation.rs/root call-sites unchanged.

Completes Phase 1: lib.rs 13,959 -> 2,033 (manifest/models/loras/recipe_presets
all extracted). Verified the only lib.rs change vs main is the two-line mod
declaration. cargo test (62 rust-api), npm run rust:check, and
cargo build --features embed-web all green.
